FORCE Fix
Some scenario where FORCE should prevail over SYNC.

+++ scripts/hhvm-repo.sh
@@ -62,7 +62,7 @@ if [[ -d $WWW_ROOT ]]; then
     echo $HHVM_MAILNAME > /etc/mailname
     service postfix start
   fi
-  if [[ -d $HHVM_SYNC_FOLDER ]] && [[ "$HHVM_SYNC" == "1" ]]; then
+  if [[ -d $HHVM_SYNC_FOLDER ]] && [[ "$HHVM_SYNC" == "1" ]] && [[ "HHVM_FORCE" == "0" ]]; then
     echo Sync $HHVM_SYNC_FOLDER to $HHVM_REPO
     mkdir -p $HHVM_REPO
     rsync -av --progress --delete $HHVM_SYNC_FOLDER/ $HHVM_REPO/
@@ -79,8 +79,11 @@ if [[ -d $WWW_ROOT ]]; then
   # Cleanup Old HHVM Data. HHVM Repo script seems to have some bug, where old data in this HHBC makes the repo build to fail.
   rm -f /root/.hhvm.hhbc
   sed -i "s/.*hhvm.repo.central.path.*/hhvm.repo.central.path=${HHVM_BC}/g" $HHVM_CFG
+  if [[ "$HHVM_FORCE" == "1" ]]; then
+    rm -f $HHVM_REPO/$REPO_FILE
+  fi
   hhvm-repo-mode enable $WWW_ROOT
-  if [[ -d $HHVM_SYNC_FOLDER ]] && [[ "$HHVM_SYNC" == "1" ]]; then
+  if [[ -d $HHVM_SYNC_FOLDER ]] && [[ "$HHVM_SYNC" == "1" ]] && [ "$HHVM_FORCE" == "0"]; then
     rsync -av --progress --delete $HHVM_REPO/ $HHVM_SYNC_FOLDER/
     rm -f $HHVM_LOCK
   fi
